requirement.txt
absl-py==0.13.0
alembic==1.7.7
altair==4.2.0
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autofeat==2.0.10
autopage==0.5.0
Babel==2.9.1
backcall==0.2.0
base58==2.1.1
black==22.1.0
bleach==4.1.0
blinker==1.4
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1636012319596/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==4.2.2
certifi==2021.10.8
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1636046173594/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1638815705608/work
clang==5.0
click==8.0.3
cliff==3.10.1
cloudpickle==2.1.0
cmaes==0.8.2
cmd2==2.4.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
colorlog==6.6.0
conda==4.12.0
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1636022092101/work
convertdate==2.4.0
creme==0.6.1
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1637648810489/work
cycler==0.11.0
Cython==0.29.28
dask==2022.5.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
descartes==1.1.0
distributed==2022.5.0
Django==4.0.4
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
featuretools==1.9.0
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
fonttools==4.29.1
fsspec==2022.5.0
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
graphviz==0.19.1
grpcio==1.39.0
h5py==3.1.0
HeapDict==1.0.1
hijri-converter==2.2.3
holidays==0.13
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1609836280497/work
importlib-metadata==4.11.3
ipykernel==6.8.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.13.4
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kneed==0.7.0
korean-lunar-calendar==0.2.1
libclang==13.0.0
llvmlite==0.38.1
locket==1.0.0
Mako==1.2.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
mizani==0.7.4
ml-metadata==0.13.1.dev0
mmh3==2.5.1
mpmath==1.2.1
msgpack==1.0.3
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numba==0.55.2
numpy==1.21.6
oauthlib==3.1.1
openpyxl==3.0.10
opt-einsum==3.3.0
optuna==2.10.0
packaging==21.3
palettable==3.3.0
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
patsy==0.5.2
pbr==5.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
Pint==0.19.2
platformdirs==2.4.1
plotnine==0.8.0
ppscore==1.1.0
prettytable==3.3.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.17.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1636020570107/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydeck==0.7.1
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.11.2
PyMeeus==0.5.11
pymongo==4.1.1
Pympler==1.0.1
pyodbc @ file:///Users/nitinverma/Downloads/pyodbc-4.0.32.tar.gz
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1633192417276/work
pyparsing==3.0.7
pyperclip==1.8.2
pyrsistent==0.18.1
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1635862736740/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
redis==4.2.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1637771257551/work
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1636009262461/work
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1640464197365/work
scikit-optimize==0.9.0
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy_1644357575279/work
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
stack-data==0.1.4
statsmodels==0.13.2
stevedore==3.5.0
streamlit==1.7.0
sympy==1.10.1
tblib==1.7.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-macos==2.7.0
tensorflow-metal==0.3.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.0
toolz==0.11.2
tornado==6.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1632160078689/work
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1632350318291/work
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.1
widgetsnbextension==3.6.0
woodwork==0.16.3
wrapt==1.12.1
xgboost==1.5.1
zict==2.2.0
zipp==3.7.0
